requirements.txt
numpy==1.23.1
streamlit==1.12.0
Image==8.2.0
tensorflow-cpu==2.9.1
protobuf==3.20.1